+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to the `C-mantic` extension will be documented in this file.
 
+## [Unreleased]
+### Fixed
+- Fixed setter generation for reference types.
+
 ## [0.4.1] - February 19, 2021
 ### Added
 - Added support for `s_` static member naming scheme when generating getters and setters.